Updates to use namespaced sandbox utilities; Prepare to split make_ursulas.
