+++ tests/network/test_network_upgrade.py
@@ -6,7 +6,7 @@ from cryptography.hazmat.primitives import serialization
 from twisted.internet import threads
 
 from nucypher.characters.lawful import Ursula
-from nucypher.utilities.sandbox import make_ursulas
+from nucypher.utilities.sandbox.ursula import make_federated_ursulas
 
 
 def test_alice_enacts_policies_in_policy_group_via_rest(enacted_federated_policy):
